+++ src/mazemaster/MazeMaster.cpp
@@ -33,8 +33,8 @@ int MazeMaster::ParseMazeArgsAndExecute(int argc, char** argv) {
 		TCLAP::CmdLine cmd("Evolved Dynamical Neural Net for solving mazes", ' ', "0.1");
 
 		//Collect all arguments
-		TCLAP::ValueArg<int> num_brains_arg("N", "num_brains", "Number of brains in each generation", true, 0, "int", cmd);
-		TCLAP::ValueArg<int> num_neurons_arg("n", "num_neurons", "Number of neuron in each brain", true, 0, "int", cmd);
+		TCLAP::ValueArg<size_t> num_brains_arg("N", "num_brains", "Number of brains in each generation", true, 0, "int", cmd);
+		TCLAP::ValueArg<size_t> num_neurons_arg("n", "num_neurons", "Number of neuron in each brain", true, 0, "int", cmd);
 
 		TCLAP::ValueArg<float> av_active_threshold_arg("a", "av_active_threshold", "Average activation threshold for each neruon", true, 0, "float", cmd);
 		TCLAP::ValueArg<float> st_dev_active_threshold_arg("A", "st_dev_active_threshold", "Standard deviation for activation threshold for each neruon", true, 0, "float", cmd);
